fix(entity-generator): fix generation of entities with identifiers that overlap with core
Accomplished by adding ts-morph as a dependency, and using it
to post process the generated sources. On a possible conflict,
user identifiers now include a marker ("U$"),
to disambiguate them in the input to ts-morph.

On conflict, the prefix "Mikro" is added to MikroORM's imports as an alias,
and if even that is not enough,
the same prefix will keep being added until the conflict is resolved.
Previously, there was a special case for the BaseEntity class.
This is now removed in favor of this more general purpose conflict resolution,
allowing an entity being named like "Entity" or "Property".

Fixed erroneously included skipped tables, and schemas.

Ensured that tables that start with a number can also be referenced properly, not merely defined,
by moving the "E" prefix addition into the naming strategy.